Change to prevent accidental removal if bindrit is unitialized

+++ uninstall.sh
@@ -2,4 +2,4 @@ bindir="pargenes/pargenes_binaries"
 rm -rf MPIScheduler/build
 rm -rf raxml-ng/build
 rm -rf modeltest/build
-rm -rf $bindir/*
+rm -rf "${bindir:?}/"*
